replace "example" with test
"example" is meant to reinforce the idea of test-as-spec, where the spec
contains examples which are written into tests. But there is so much
cultural momentum that I can't even remember to use the word "example"
in my own head. So I resign myself to the use of the word "test".